+++ database/migrations/2018_09_29_212046_add_groupflight_to_event_types_table.php
@@ -1,0 +1,28 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+
+class AddGroupflightToEventTypesTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        DB::table('event_types')->insert([
+            'id' => 4, 'name' => 'Groupflight'
+        ]);
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        DB::table('event_types')->delete(4);
+    }
+}